[BRR-ICE ERROR]  Fixed bug in logic of magento re-login retry
On brr-ice.com, we noticed that the magento login retry was not working
properly, and we encountered the following exception:

```
***** Session re-login successful.
TypeError: undefined is not a function
    at Object.exports.serializeMethodCall (/var/app/current/node_modules/magento-api/node_modules/xmlrpc/lib/serializer.js:25:10)
    at Client.methodCall (/var/app/current/node_modules/magento-api/node_modules/xmlrpc/lib/client.js:100:30)
    at Magento.methodApply (/var/app/current/node_modules/magento-api/src/magento.js:158:15)
    at /var/app/current/node_modules/magento-api/src/magento.js:180:28
    at /var/app/current/node_modules/magento-api/src/magento.js:205:5
    at callback (/var/app/current/node_modules/magento-api/node_modules/xmlrpc/lib/deserializer.js:42:7)
    at Deserializer.onDone (/var/app/current/node_modules/magento-api/node_modules/xmlrpc/lib/deserializer.js:92:12)
    at SAXStream.emit (events.js:104:17)
    at Object.SAXStream._parser.onend (/var/app/current/node_modules/magento-api/node_modules/xmlrpc/node_modules/sax/lib/sax.js:171:8)
    at emit (/var/app/current/node_modules/magento-api/node_modules/xmlrpc/node_modules/sax/lib/sax.js:325:33)
Connecting to Redis at: dev-re-1t8rhc3umrq86.laaefq.0001.use1.cache.amazonaws.com
```

Based on this exception trace, we thought it most likely that there was a bug
in the logic of our magento timeout handler.  This commit attempts to fix that
error by creating a copy of the arguments object earlier in the method so that
the nature of the nested callback fuctions we are using does not obscure the
aet of `arguments` that are needed for the retry call.